// gen/ts-fetch/models/WebhookSubscriptionResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WebhookSubscriptionResponse = {
    webhookId?: string;
    resellerId?: string | null;
    customerId?: string | null;
    url?: string;
    secret?: string;
    eventTypes?: Array<string>;
    enabled?: boolean;
    description?: string | null;
    createdAt?: string;
    updatedAt?: string;
};