// gen/ts-fetch/models/WebhookSubscriptionUpdateRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WebhookSubscriptionUpdateRequest = {
    url?: string;
    secret?: string;
    eventTypes?: Array<string>;
    enabled?: boolean;
    description?: string | null;
};